requirements.txt
torch
torchvision
numpy
pandas
matplotlib
tqdm
Pillow
